+++ src/App.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Home from "./routes/Home"
-import About from "./routes/About"
-import Contact from "./routes/Contact"
-import Experience from "./routes/Experience"
+import Home from "./components/routes/Home"
+import About from "./components/routes/About"
+import Contact from "./components/routes/Contact"
+import Experience from "./components/routes/Experience"
 
 function App() {
   return (
