+++ src/components/Dropdown.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import styles from "../css/Dropdown"
+import styles from "../css/Dropdown.module.css"
 
 const Dropdown = () => {
