+++ src/components/routes/About.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import styles from "../css/About.module.css"
+import styles from "../../css/About.module.css"
 
 const About = () => {
