+++ src/components/routes/Experience.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import styles from "../css/Experience.module.css"
+import styles from "../../css/Experience.module.css"
 
 const Experience = () => {
